Create a custom python toolchain based on the FreeCAD binary.
The FreeCAD's AppImage contains a custom version of python,
packaged with the pip dependencies and compiled libraries required to
run FreeCAD.

This commit defines a custom toolchain in Bazel,
so the FreeCAD binary can be used in headless mode as a python
interpreter,
which enables the users to provide a FreeCAD macro in a .py file
and run it with the freecadcmd within the FreeCAD AppImage.

This commit is a proof of concept and
focuses on on linux_86_64 systems.
The work to enable other operating systems and other CPU architectures
is left for future work.

The toolchain that this commit create is hermetic,
i.e. the toolchain does not depend on specific libraries or dependencies
on the system.
Any dependency required by the toolchain in downloaded and verified
(using sha256 hashes).

Bazel's built in cache prevents downloading the AppImage on consecutive
builds.