Fix the labels in rules_freecad
Before the fix, anether bazel project could not use rules_freecad
because when rules_freecad was used as a dependency,
the extension implementation hardcoded label strings
that are only valid when rules_freecad is the root module.

The fix uses  Label() in the Starlark implementation
to resolve labels to their canonical forms which are valid globally.
- e.g., @@rules_freecad~...//...